backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jackson.wen <jackson.wen@xm.wonder.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
fastapi = {extras = ["all"], version = "^0.104.1"}
redis = "^5.0.1"
celery = "^5.3.4"
asgi-request-id = "^0.1.2"
apscheduler = "^3.10.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.1"
slowapi = "^0.1.8"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
mypy = "^1.6.1"
black = "^23.10.1"
isort = "^5.12.0"
autoflake = "^2.2.1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
